ENH: split execution spaces in separate files
This commit, in part, resolves #56
